requirements.txt
flask==2.2.3
flask_cors==3.0.1
python-dotenv==1.0.0
requests==2.28.2
zenpy==2.0.25